pipeline/requirements.txt
azure-functions==1.21.3
pymssql==2.3.2
